requirements.txt
django==5.0.3
channels==4.2.2
aiosignal
pillow==10.4.0
asgiref==3.8.1
whitenoise
daphne
Pillow
